Payment methods page

+++ src/app/controllers/settings/methods.js
@@ -2,8 +2,11 @@ angular.module('proton.controllers.Settings')
 
 .controller('MethodsController', function(
     $scope,
+    $translate,
     cardModal,
-    methods
+    confirmModal,
+    methods,
+    notify
 ) {
     $scope.methods = methods.data.PaymentMethods;
 
@@ -29,7 +32,21 @@ angular.module('proton.controllers.Settings')
     };
 
     $scope.delete = function(method) {
-        Payment.deleteMethod(method.ID);
-        $scope.methods.splice($scope.methods.indexOf(method), 1);
+        confirmModal.activate({
+            params: {
+                title: $translate.instant('DELETE_PAYMENT_METHOD'),
+                message: $translate.instant('Are you sure you want to delete this payment method?'),
+                confirm: function() {
+                    Payment.deleteMethod(method.ID)
+                    .then(function(result) {
+                        if (result.data && result.data.Code === 1000) {
+                            confirmModal.deactivate();
+                            $scope.methods.splice($scope.methods.indexOf(method), 1);
+                            notify({message: $translate.instant('PAYMENT_METHOD_DELETED'), classes: 'notification-danger'});
+                        }
+                    });
+                }
+            }
+        });
     };
 });
